feat(exception): modify type of a parameter
Money 값 객체의 값이 long 타입으로 변경됨에 따라 예외에서 받을 파라미터의 타입을 변경했습니다.

+++ src/main/java/lotto/common/exception/InvalidMoneyException.java
@@ -2,11 +2,11 @@ package lotto.common.exception;
 
 public class InvalidMoneyException extends LottoException{
 
-    public InvalidMoneyException(int money) {
+    public InvalidMoneyException(long money) {
         super(ErrorMessage.MONEY_OUT_OF_RANGE.getMessage() + "(금액: " + money + ")");
     }
 
-    public InvalidMoneyException(int money, Exception e) {
+    public InvalidMoneyException(long money, Exception e) {
         super(ErrorMessage.MONEY_OUT_OF_RANGE.getMessage() + "(금액: " + money + ")", e);
     }
 }
